:recycle: Refactor: Double를 확장, CalculateItem 채택.
Double을 확장시켜 CalculateItem을 채택하게 만들었습니다.

+++ Calculator/Calculator/Model/CalculatorItemQueue.swift
@@ -5,7 +5,7 @@
 //  Created by Minseong Kang on 2023/05/30.
 //
 
-extension Int: CalculateItem { }
+extension Double: CalculateItem { }
 
 protocol CalculateItem: Equatable { }
